pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "chatroom"
version = "0.1.0"
description = "Encrypted TCP chatroom (RSA key exchange + AES-EAX framed messages + TOFU)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pycryptodome",
]

[project.scripts]
chatroom = "chatroom.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

